+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 - added `quit()` to end everything
 - shader error logs now yields the correct line number
 - changed object update order from reversed to not reversed
+- added `download()`, `downloadText()`, `downloadJSON()`, `downloadBlob()`
+- added `Recording#stop()` to stop the recording and returns the video data as mp4 Blob
 
 ### v2000.2.6